+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/filter/BinaryComparison.java
@@ -20,36 +20,41 @@ package org.apache.spark.sql.connector.expressions.filter;
 import java.util.Objects;
 
 import org.apache.spark.annotation.Evolving;
+import org.apache.spark.sql.connector.expressions.Literal;
 import org.apache.spark.sql.connector.expressions.NamedReference;
 
 /**
- * Base class for {@link IsNull}, {@link IsNotNull}
+ * Base class for {@link EqualNullSafe}, {@link EqualTo}, {@link GreaterThan},
+ * {@link GreaterThanOrEqual}, {@link LessThan}, {@link LessThanOrEqual}
  *
  * @since 3.3.0
  */
 @Evolving
-public class NullCmpFilter extends Filter {
+public abstract class BinaryComparison extends Filter {
   protected final NamedReference column;
+  protected final Literal<?> value;
 
-  protected NullCmpFilter(NamedReference column) {
+  protected BinaryComparison(NamedReference column, Literal<?> value) {
     this.column = column;
+    this.value = value;
   }
 
   public NamedReference column() { return column; }
+  public Literal<?> value() { return value; }
 
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
-    NullCmpFilter isNull = (NullCmpFilter) o;
-    return Objects.equals(column, isNull.column);
+    BinaryComparison that = (BinaryComparison) o;
+    return Objects.equals(column, that.column) && Objects.equals(value, that.value);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(column);
+    return Objects.hash(column, value);
   }
 
   @Override
-  public NamedReference[] references() { return new NamedReference[]{column}; }
+  public NamedReference[] references() { return new NamedReference[] { column }; }
 }
