+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/filter/EqualNullSafe.java
@@ -29,7 +29,7 @@ import org.apache.spark.sql.connector.expressions.NamedReference;
  * @since 3.3.0
  */
 @Evolving
-public final class EqualNullSafe extends BinaryFilter {
+public final class EqualNullSafe extends BinaryComparison {
 
   public EqualNullSafe(NamedReference column, Literal<?> value) {
     super(column, value);
