+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/filter/GreaterThan.java
@@ -28,7 +28,7 @@ import org.apache.spark.sql.connector.expressions.NamedReference;
  * @since 3.3.0
  */
 @Evolving
-public final class GreaterThan  extends BinaryFilter {
+public final class GreaterThan  extends BinaryComparison {
 
   public GreaterThan(NamedReference column, Literal<?> value) {
     super(column, value);
