+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/filter/In.java
@@ -71,5 +71,5 @@ public final class In extends Filter {
   }
 
   @Override
-  public NamedReference[] references() { return new NamedReference[]{column}; }
+  public NamedReference[] references() { return new NamedReference[] { column }; }
 }
