+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/filter/StringEndsWith.java
@@ -28,7 +28,7 @@ import org.apache.spark.unsafe.types.UTF8String;
  * @since 3.3.0
  */
 @Evolving
-public final class StringEndsWith extends StringCmpFilter {
+public final class StringEndsWith extends StringPredicate {
 
   public StringEndsWith(NamedReference column, UTF8String value) {
     super(column, value);
